+++ src/ts/utils/helpers/BezierPathInteractionHelper.ts
@@ -27,6 +27,7 @@
  * @modified 2025-04-14 BezierPathInteractionHelper: Changed default value of `HelperOptions.autoAdjustPaths` from `true` to `false`.
  * @modified 2025-05-05 Added optional params `draw` and `fill` to BezierPathInteractionHelper.drawHandleLines` method.
  * @modified 2025-05-05 Class `BezierPathInteractionHelper` now implementing `IShapeInteractionHelper`.
+ * @modified 2025-05-07 Tweaking performance of `BezierPathInteractionHelper`: only triggering redraw now when mouse move within given detect range.
  * @version  1.2.1
  *
  * @file BezierPathInteractionHelper
@@ -440,6 +441,9 @@ export class BezierPathInteractionHelper implements IShapeInteractionHelper {
   // | Touch and mouse events should call this fuction when the pointer was moved.
   // +-------------------------------
   private _handleMoveEvent(posX: number, posY: number): void {
+    const oldA = this.currentA.clone();
+    const oldB = this.currentB.clone();
+    const oldDistance = this.currentDistance;
     const point: XYCoords = this.pb.transformMousePosition(posX, posY);
     this.currentB.set(point);
     this._updateMinDistance();
@@ -449,8 +453,14 @@ export class BezierPathInteractionHelper implements IShapeInteractionHelper {
     } else {
       this.onPointerMoved(-1, null, null, 0.0);
     }
-    // Always redraw even when moving outside the detection distance?
-    this.pb.redraw();
+    // Only redraw when moving inside the detection distance
+    if (
+      (this.currentDistance <= this.maxDetectDistance ||
+        (this.currentDistance > this.maxDetectDistance && oldDistance <= this.maxDetectDistance)) &&
+      (!oldA.equals(this.currentA) || !oldB.equals(this.currentB))
+    ) {
+      this.pb.redraw();
+    }
   }
 
   // +---------------------------------------------------------------------------------
@@ -468,23 +478,15 @@ export class BezierPathInteractionHelper implements IShapeInteractionHelper {
   private _installTouchListener(): AlloyFinger {
     var _self: BezierPathInteractionHelper = this;
     const afProps = {
-      // Todo: which event types does AlloyFinger use?
-      // touchStart: function (_event: TouchEvent) {
-      //   _self.mouseIsOver = true;
-      // },
-      // /* TouchMoveEvent , AFTouchEvent<"touchStart"> */
       touchStart: function (_event: AFTouchEvent<"touchStart">) {
         _self.mouseIsOver = true;
       },
       touchMove: function (event: AFTouchEvent<"touchMove">) {
-        // TouchEvent) {
         if (_self.pb.getDraggedElementCount() == 0 && event.touches.length > 0) {
-          // console.log('touchmove');
           _self._handleMoveEvent(event.touches[0].clientX, event.touches[0].clientY);
         }
       },
       touchEnd: function (_event: AFTouchEvent<"touchEnd">) {
-        // TouchEvent) {
         _self.mouseIsOver = false;
         _self._clearMoveEvent();
       }
@@ -577,7 +579,9 @@ export class BezierPathInteractionHelper implements IShapeInteractionHelper {
   // | Update the min distance from point `line.b` to the curve. And redraw.
   // +-------------------------------
   private _updateMinDistance(): void {
-    if (this.paths.length == 0) return;
+    if (this.paths.length == 0) {
+      return;
+    }
     let pathIndex: number = -1;
     let minDist: number = Number.MAX_VALUE;
     let closestPoint: Vertex | null = null;